+++ arch/ppc64/include/types.h
@@ -38,8 +38,8 @@ typedef signed long long __s64;
 
 typedef unsigned char __u8;
 typedef unsigned short __u16;
-typedef unsigned long __u32;
-typedef unsigned long long __u64;
+typedef unsigned int __u32;
+typedef unsigned long __u64;
 
 typedef __u64 __address;
 typedef __u32 pfn_t;
